UPDATE: v0.8.0 Recode
Full of bugs. I am aware. Currently flashing screen does not work, alert sound does not work, and it still does not work in the background. It is still a client side application. I am still trying to decipher the v0.8.0 DeskThing recode. I have another version that should solve all of these issues once I get it working. But for now I wanted to release a temporary semi functional app so that it is available while I finish working on it. I also understand this app breaks button mapping... You will need to swipe to navigate apps. I apologize for this. A lot of the v0.8.0 recode is unfamiliar to me and Im having to do a lot of learning to get this even semi working lol. But enjoy the stone-age app